SQL: Fixed Out of range currentBid at characters database in table blackmarket_auctions.

+++ sql/base/characters_database.sql
@@ -312,7 +312,7 @@ CREATE TABLE `blackmarket_auctions` (
   `templateId` int(10) unsigned NOT NULL DEFAULT '0',
   `startTime` int(10) unsigned NOT NULL DEFAULT '0',
   `currentBidder` int(10) unsigned NOT NULL DEFAULT '0',
-  `currentBid` int(10) unsigned NOT NULL DEFAULT '0',
+  `currentBid` bigint(20) unsigned NOT NULL DEFAULT '0',
   `minIncrement` int(10) unsigned NOT NULL DEFAULT '0',
   `numBids` int(10) unsigned NOT NULL DEFAULT '0',
   PRIMARY KEY (`auctionId`),
